First draft of git parsing
